==> 2 Addestramento AE.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

%cd drive/MyDrive/Tesi\ Magistrale\ 204963/Progetto/

import warnings
warnings.filterwarnings("ignore")

YAHOO
===

!python autoencoders/train.py --train data/yahoo/train_2_pre.txt --valid data/yahoo/test_2_pre.txt --model_type aae --lambda_adv 10 --noise 0.3,0,0,0 --batch-size 512 --vocab-size 25000 --save-dir mytest/checkpoint/yahoo/daae_1

!python autoencoders/train.py --load-model mytest/checkpoint/yahoo/daae_1/model.pt --train data/yahoo/train_2_pre.txt --valid data/yahoo/test_2_pre.txt --model_type aae --lambda_adv 10 --noise 0.3,0,0,0 --batch-size 512 --vocab-size 25000 --save-dir mytest/checkpoint/yahoo/daae_1

!python autoencoders/test.py --sample --n 100 --checkpoint mytest/checkpoint/yahoo/daae_1/ --output sample.txt

with open("mytest/checkpoint/yahoo/daae_1/sample.txt",'r') as f:
  data = f.readlines()
  f.close()

data

!python autoencoders/test.py --reconstruct --data data/yahoo/test_2_pre.txt --output test_2.rec --checkpoint mytest/checkpoint/yahoo/daae_1/

YELP
===

!python autoencoders/train.py --train data/yelp/train.txt --valid data/yelp/valid.txt --model_type aae --lambda_adv 10 --noise 0.3,0,0,0 --batch-size 512 --vocab-size 12000 --save-dir mytest/checkpoint/yelp/daae_1

TWITTER
===

!python autoencoders/train.py --train data/twitter/twitter_train.txt --valid data/twitter/twitter_test.txt --model_type aae --lambda_adv 10 --noise 0.3,0,0,0 --batch-size 512 --vocab-size 12000 --save-dir mytest/checkpoint/twitter/daae_1
